+++ src/self_similarity.py
